pyproject.toml
[tool.poetry]
name = "livex-ai-challenge"
version = "0.1.0"
description = ""
authors = ["Bharath Sriraam R R <albharath.1305@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.35.7"
fastapi = "0.110.1"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
ipykernel = "^6.29.5"
chainlit = "^1.1.305"
uvicorn = {version = "0.25.0", extras = ["standard"]}
instructor = "^1.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
